+++ src/main/java/dinkplugin/util/MathUtils.java
@@ -10,6 +10,14 @@ public class MathUtils {
     public static final double EPSILON = 0.00001;
     private static final int[] FACTORIALS;
 
+    public int sum(int[] array) {
+        int x = 0;
+        for (int i : array) {
+            x += i;
+        }
+        return x;
+    }
+
     public String formatPercentage(double d, int sigFigs) {
         return BigDecimal.valueOf(d * 100)
             .round(new MathContext(sigFigs))